feat: defitrack-makerdao: Move lending protocol implementation to makerdao folder.
Ref: https://github.com/defitrack/defitrack-core/issues/16
